backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.4
rq==1.16.2
rq-scheduler==0.13.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
aiofiles==23.2.1
prometheus-client==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
tiktoken==0.7.0
sentence-transformers==2.7.0
numpy==1.26.4
# Ray Serve (optional, comment out if not using)
# ray[serve]==2.9.0
# vllm==0.4.2
